+++ app/controllers/comments_controller.rb
@@ -6,7 +6,10 @@ class CommentsController < ApplicationController
   rescue_from ActiveRecord::RecordNotFound, with: :render_404
 
   def create
-    ParticipationFactory.create(task: comment.task, assignee: current_user, assigner: current_user)
+    if !current_user.journal_admin? journal: task.paper.journal
+      ParticipationFactory.create(task: comment.task, assignee: current_user, assigner: current_user)
+    end
+    Activity.comment_created! comment, user: current_user
     respond_with comment if CommentLookManager.sync_comment(comment)
   end