+++ app/models/paper.rb
@@ -138,6 +138,12 @@ class Paper < ActiveRecord::Base
     end
   end
 
+  def download_body
+    if body
+      "#{body}#{download_supporting_information}"
+    end
+  end
+
   def version_string
     latest_version.version_string
   end
@@ -326,5 +332,19 @@ class Paper < ActiveRecord::Base
   def set_submitting_user_and_touch!(submitting_user) # rubocop:disable Style/AccessorMethodName
     latest_version.update!(submitting_user: submitting_user)
     latest_version.touch
+  end
+
+  def download_supporting_information
+    return if supporting_information_files.empty?
+
+    supporting_information = "<h2>Supporting Information</h2>"
+    supporting_information_files.each do |file|
+      if file.preview_src
+        supporting_information.concat "<p>#{file.download_link file.preview_image}</p>"
+      end
+      supporting_information.concat "<p>#{file.download_link}</p>"
+    end
+
+    supporting_information
   end
 end