+++ engines/plos_authors/app/models/plos_authors/plos_author.rb
@@ -1,7 +1,7 @@
 module PlosAuthors
   class PlosAuthor < ActiveRecord::Base
-  include EventStream::Notifiable
-  include MetadataTask
+    include EventStream::Notifiable
+    include MetadataTask
 
     acts_as :author, dependent: :destroy
     delegate :completed?, to: :plos_authors_task, prefix: :task, allow_nil: true